* v1.x: (271 commits)
  flush before rotating
  Throw if rotating_file_sink constructor receives max_size==0 as arg
  comment
  Fix #2261
  throw if flush failed
  added a few missing files/directories to the gitignore
  removed unneeded spaces
  Update pattern_formatter-inl.h
  Added the last few suggested changes.
  Several minor improvements based on code review suggestions.
  Fix mistake in previous checkin.
  Resetting the needs time flag when setting a pattern.
  Now only getting the date if formater needs to display date related information.
  Allow forward-declaration of level_enum
  Reset current size if rotated files on open
  spdlog: fmt - support `std::span` in `to_hex`
  🚧 fix building with c++11
  🔥 conditional_cast
  Revert "Useless cast"
  Useless cast
  ...

+++ include/spdlog/sinks/ansicolor_sink-inl.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #ifndef SPDLOG_HEADER_ONLY
-#include <spdlog/sinks/ansicolor_sink.h>
+#    include <spdlog/sinks/ansicolor_sink.h>
 #endif
 
 #include <spdlog/pattern_formatter.h>
@@ -34,7 +34,7 @@ template<typename ConsoleMutex>
 SPDLOG_INLINE void ansicolor_sink<ConsoleMutex>::set_color(level::level_enum color_level, string_view_t color)
 {
     std::lock_guard<mutex_t> lock(mutex_);
-    colors_[color_level] = to_string_(color);
+    colors_[static_cast<size_t>(color_level)] = to_string_(color);
 }
 
 template<typename ConsoleMutex>
@@ -52,7 +52,7 @@ SPDLOG_INLINE void ansicolor_sink<ConsoleMutex>::log(const details::log_msg &msg
         // before color range
         print_range_(formatted, 0, msg.color_range_start);
         // in color range
-        print_ccode_(colors_[msg.level]);
+        print_ccode_(colors_[static_cast<size_t>(msg.level)]);
         print_range_(formatted, msg.color_range_start, msg.color_range_end);
         print_ccode_(reset);
         // after color range
